# api/openapi.json
openapi: 3.0.0
info:
  title: SlayFlashcards API
  description: REST API for SlayFlashcards - An interactive flashcard learning platform
  version: 1.0.0
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /flashcards:
    get:
      summary: Get flashcards
      tags:
        - flashcards
      description: Get flashcards with optional filtering and pagination.
      security:
        - bearerAuth: []
      parameters:
        - name: quiz_id
          in: query
          description: Filter by quiz ID
          schema:
            type: integer
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: difficulty
          in: query
          description: Filter by difficulty
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: question_lang
          in: query
          description: Filter by question language
          schema:
            type: string
        - name: answer_lang
          in: query
          description: Filter by answer language
          schema:
            type: string
        - name: answer_type
          in: query
          description: Filter by answer type
          schema:
            $ref: '#/components/schemas/AnswerType'
        - name: search_text
          in: query
          description: Search by question or answer text
          schema:
            type: string
      responses:
        '200':
          description: Retrieved list of flashcards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlashcardsResponse'
        '400':
          description: Invalid or missing parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new flashcard
      tags:
        - flashcards
      description: Create a new flashcard.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlashcardCreate'
      responses:
        '201':
          description: Flashcard created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlashcardResponse'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /flashcards/{flashcard_id}:
    get:
      summary: Get a specific flashcard by ID
      tags:
        - flashcards
      description: Get a specific flashcard by ID.
      security:
        - bearerAuth: []
      parameters:
        - name: flashcard_id
          in: path
          required: true
          description: The ID of the flashcard to retrieve.
          schema:
            type: integer
      responses:
        '200':
          description: Flashcard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlashcardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Flashcard not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an existing flashcard
      tags:
        - flashcards
      description: Update an existing flashcard.
      security:
        - bearerAuth: []
      parameters:
        - name: flashcard_id
          in: path
          required: true
          description: The ID of the flashcard to update.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlashcardUpdate'
      responses:
        '200':
          description: Flashcard updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlashcardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Flashcard not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a flashcard
      tags:
        - flashcards
      description: Delete a flashcard.
      security:
        - bearerAuth: []
      parameters:
        - name: flashcard_id
          in: path
          required: true
          description: The ID of the flashcard to delete.
          schema:
            type: integer
      responses:
        '204':
          description: Flashcard deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Flashcard not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /flashcards/bulk:
    post:
      summary: Create multiple flashcards at once
      tags:
        - flashcards
      description: Create multiple flashcards at once.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkFlashcardCreate'
      responses:
        '201':
          description: Bulk operation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResponse'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /flashcards/quiz/{quiz_id}/random:
    get:
      summary: Get random flashcards from a quiz
      tags:
        - flashcards
      description: Get random flashcards from a quiz.
      security:
        - bearerAuth: []
      parameters:
        - name: quiz_id
          in: path
          required: true
          description: The ID of the quiz to get flashcards from.
          schema:
            type: integer
        - name: count
          in: query
          description: Number of random flashcards
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 50
        - name: difficulty
          in: query
          description: Filter by difficulty
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: Retrieved list of random flashcards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlashcardsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /flashcards/quiz/{quiz_id}/types:
    get:
      summary: Get answer type statistics for flashcards in a quiz
      tags:
        - flashcards
      description: Get answer type statistics for flashcards in a quiz.
      security:
        - bearerAuth: []
      parameters:
        - name: quiz_id
          in: path
          required: true
          description: The ID of the quiz to get statistics for.
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved answer type statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total_flashcards:
                        type: integer
                      answer_types:
                        type: object
                        additionalProperties:
                          type: integer
                      available_types:
                        type: array
                        items:
                          type: string
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnswerType:
      type: string
      enum:
        - text
        - integer
        - float
        - range
        - boolean
        - choice
        - multiple_choice
        - short_text
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    FlashcardQuestionBase:
      type: object
      required:
        - title
        - text
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        text:
          type: string
          minLength: 1
          maxLength: 2000
        lang:
          type: string
          maxLength: 10
          nullable: true
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        emoji:
          type: string
          maxLength: 10
          nullable: true
        image:
          type: string
          maxLength: 500
          nullable: true
    FlashcardAnswerBase:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
        lang:
          type: string
          maxLength: 10
          nullable: true
        type:
          $ref: '#/components/schemas/AnswerType'
          default: text
        options:
          type: array
          items:
            type: object
          nullable: true
        metadata:
          type: object
          nullable: true
    FlashcardBase:
      type: object
      required:
        - quiz_id
        - question
        - answer
      properties:
        quiz_id:
          type: integer
        question:
          $ref: '#/components/schemas/FlashcardQuestionBase'
        answer:
          $ref: '#/components/schemas/FlashcardAnswerBase'
    FlashcardCreate:
      allOf:
        - $ref: '#/components/schemas/FlashcardBase'
    FlashcardUpdate:
      type: object
      properties:
        question:
          $ref: '#/components/schemas/FlashcardQuestionBase'
        answer:
          $ref: '#/components/schemas/FlashcardAnswerBase'
    Flashcard:
      type: object
      properties:
        id:
          type: integer
        quiz_id:
          type: integer
        question_title:
          type: string
        question_text:
          type: string
        question_lang:
          type: string
          nullable: true
        question_difficulty:
          type: integer
          nullable: true
        question_emoji:
          type: string
          nullable: true
        question_image:
          type: string
          nullable: true
        answer_text:
          type: string
        answer_lang:
          type: string
          nullable: true
        answer_type:
          type: string
        answer_options:
          type: array
          items:
            type: object
          nullable: true
        answer_metadata:
          type: object
          nullable: true
    FlashcardResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Flashcard'
    FlashcardsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Flashcard'
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          nullable: true
        message:
          type: string
        code:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          nullable: true
        request_id:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    BulkFlashcardCreate:
      type: object
      required:
        - quiz_id
        - flashcards
      properties:
        quiz_id:
          type: integer
        flashcards:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            properties:
              question:
                type: object
                properties:
                  title:
                    type: string
                  text:
                    type: string
                  lang:
                    type: string
                  difficulty:
                    type: integer
                  emoji:
                    type: string
                  image:
                    type: string
              answer:
                type: object
                properties:
                  text:
                    type: string
                  lang:
                    type: string
                  type:
                    $ref: '#/components/schemas/AnswerType'
                  options:
                    type: array
                    items:
                      type: object
                  metadata:
                    type: object
    BulkOperationResult:
      type: object
      required:
        - total
        - successful
        - failed
      properties:
        total:
          type: integer
        successful:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
    BulkOperationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BulkOperationResult'